fix: keyword 검색 조회 결과 null 버그 수정

+++ src/main/java/com/capstone/mall/service/category/CategoryServiceImpl.java
@@ -76,6 +76,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public ResponseDto deleteCategory(Long categoryId) {
+        if (categoryRepository.existsById(categoryId)) {
+            return responseService.createResponseDto(200, "Category does not exist", null);
+        }
+
         categoryRepository.deleteById(categoryId);
 
         return responseService.createResponseDto(200, "", categoryId);
